requirements.txt
google-cloud-bigquery==3.34.0
google-cloud-logging==3.12.1
google-cloud-storage==2.19.0
google-cloud-secret-manager==2.23.3
openpyxl==3.1.5
pandas==2.2.3